requirements.txt
streamlit
requests
faiss-cpu
sentence-transformers
numpy
yt-dlp
python-dotenv
